Gives Factory Refills names
Changes the name of all factory refills to not all just be "rounded metal plates", instead gives them a suffix [(AT APHE), for example] to tell which item this refill will produce, as well as update the description to reflect the item this refill makes. Entries that had a unique name are unchanged.