Introduce a factorio valhalla vendor
Introduce a factorio valhalla vendor that contains all but one req refills (the one exception being drop pod as I believe that could be used to deploy on the map). For the purpose of trying out factorio designs. All placed one vendor in the westernmost part of the valhalla as I could not figure where to make a proper factorio room yet.