Converted SAAspeedloaders to new class
Converted the SAA speedloaders to their own class that matches the class of the SAA revolver.  Created an improper class as a gap, given there is also an improper class for "single action" revolvers, even though there is only one in the game.

Replaced all instances of /obj/item/ammo_magazine/revolver with the new speedloader