Malfunction() override
The parent malfunction() proc breaks the vending machine when called, which is unsuitable for the nanoammo so it was removed.  The parent proc also vends items linearly on each malfunction, which is fine for other vending machines which break on the first malfunction, but the nanoammo is intended to survive a malfunction.  Therefore, this loop ensures a random item is deleted instead of the first.  Also, it deletes items instead of dispensing them to reduce lag because repeated dispenses from malfunctions were causing a significant lag.