Overrides attack_alien() proc
The machinery/vending/attack_alien() proc does not call the parent /obj/attack_alien() proc, which means many aspects are unavailable to child vending machines, such as taking damage at all.  This is a snowflake solution to get the intended functionality of allowing the nanoammo to be damaged by slashes while disabling the ability to tip the nanoammo (as its wallmounted and has no reason to tilt).  However, this solution is entirely a bandaid solution to the real problem which is the problematic machinery/vending/attack_alien() proc.  Solving this, however, is out of scope for this PR.  Ultimately, this commit can be discarded without any significant issues to the function of the nanoammo, it just means that xenos can only destroy the nanoammo by aciding it.