Simplified attack_alien() override
Previous override was too focused on preventing tipping, given the NanoAmmo is wall mounted and doesn't necessarily tip.  However, the solution to that is best solved with an addition of a flag that can be added with a different PR.

It was better for everything to just call the parent proc and add the attack_generic proc to allow the NanoAmmo to take damage.